+++ modules/linux-function/locals.tf
@@ -1,6 +1,8 @@
 locals {
+  is_consumption = data.azurerm_service_plan.plan.sku_name == "Y1"
+
   default_site_config = {
-    always_on                              = data.azurerm_service_plan.plan.sku_name == "Y1" ? false : true
+    always_on                              = !local.is_consumption
     application_insights_connection_string = var.application_insights_enabled ? local.app_insights.connection_string : null
     application_insights_key               = var.application_insights_enabled ? local.app_insights.instrumentation_key : null
   }